feat: :tada: (index.html): Init
This is an HTML document that appears to be a basic to-do list application. The document includes a form for adding a new to-do item, and a list group displaying existing to-do items. The to-do items are displayed as list items with checkboxes and a trash icon for deleting items. There are also buttons for filtering the list by "all", "todo", and "done" items. The document also includes links to the Bootstrap and Bootstrap-icons CSS libraries for styling.